feat: Add basic HTTP server functionality
The code changes in `server.js` add basic HTTP server functionality using the `createServer` function from the `http` module. This allows the server to respond with a "Hello, world!" message when a request is made.

Note: The suggested commit message follows the established convention of using a prefix (feat) to indicate a new feature or functionality being added.